+++ MainCore/Errors/MustRetry.cs
@@ -1,0 +1,11 @@
+using FluentResults;
+
+namespace MainCore.Errors
+{
+    public class MustRetry : Error
+    {
+        public MustRetry(string message) : base($"{message}. Bot must retry")
+        {
+        }
+    }
+}
