+++ MainCore/Errors/NeedLogin.cs
@@ -1,0 +1,11 @@
+﻿using FluentResults;
+
+namespace MainCore.Errors
+{
+    public class NeedLogin : Error
+    {
+        public NeedLogin() : base("Account is logged out.")
+        {
+        }
+    }
+}
