+++ MainCore/Errors/NetworkProblem.cs
@@ -1,0 +1,11 @@
+﻿using FluentResults;
+
+namespace MainCore.Errors
+{
+    public class NetworkProblem : Error
+    {
+        public NetworkProblem(string message) : base($"{message}.")
+        {
+        }
+    }
+}
