--- MainCore/Exceptions/ChromeMissingException.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace MainCore.Exceptions
-{
-    public class ChromeMissingException : Exception
-    {
-    }
-}
