--- MainCore/Exceptions/LoginNeedException.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace MainCore.Exceptions
-{
-    public class LoginNeedException : Exception
-    {
-    }
-}
