--- MainCore/Exceptions/StopNowException.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace MainCore.Exceptions
-{
-    public class StopNowException : Exception
-    {
-        public StopNowException(string message) : base(message)
-        {
-        }
-    }
-}
