+++ MainCore/Helper/Interface/IAccessHelper.cs
@@ -1,0 +1,9 @@
+using RestSharp;
+
+namespace MainCore.Helper.Interface
+{
+    public interface IAccessHelper
+    {
+        bool IsValid(RestClient client);
+    }
+}
