+++ MainCore/Helper/Interface/IClickHelper.cs
@@ -1,0 +1,13 @@
+﻿using FluentResults;
+
+namespace MainCore.Helper.Interface
+{
+    public interface IClickHelper
+    {
+        Result ClickCompleteNow(int accountId);
+
+        Result ClickStartAdventure(int accountId, int x, int y);
+
+        Result ClickStartFarm(int accountId, int farmId);
+    }
+}
